backend/requirements.txt
django>=5.0,<6.0
djangorestframework>=3.15
psycopg[binary]>=3.2
python-dotenv>=1.0
django-cors-headers>=4.4
django-filter>=24.3
djangorestframework-simplejwt>=5.3
pytest>=8.2
pytest-django>=4.9